fix: polish release prep for v0.1.0
- Add pip cache to CI workflow for faster runs
- Remove bedrock/vertex optional deps (not implemented)
- Extract shared mock provider to examples/_mock.py
- Fix unnecessary f-string in example 03
